Extracted toolbar from text editor component.
