+++ client/src/components/index.js
@@ -1,5 +1,5 @@
 import SignUpContainer from "./SignUpContainer";
 import ErrorMessage from "./ErrorMessage";
-import TextEditor from "./TextEditor";
+import TextEditor from "./TextEditor/TextEditor";
 
 export { SignUpContainer, ErrorMessage, TextEditor };
